+++ packages/@expo/config-plugins/src/utils/modules.ts
@@ -1,5 +1,7 @@
 import fs from 'fs';
 
+const debug = require('debug')('expo:config-plugins:fs') as typeof console.log;
+
 /**
  * A non-failing version of async FS stat.
  *
@@ -27,4 +29,16 @@ export function fileExists(file: string): boolean {
   } catch {
     return false;
   }
+}
+
+// An optimization to attempt to prevent Xcode cache invalidation on files that don't change.
+export async function writeIfDifferentAsync(filePath: string, contents: string): Promise<void> {
+  if (!fileExists(filePath)) {
+    const existing = await fs.promises.readFile(filePath, 'utf8');
+    if (existing === contents) {
+      debug(`Skipping writing unchanged file: ${filePath}`);
+      return;
+    }
+  }
+  await fs.promises.writeFile(filePath, contents);
 }